added messages template

+++ app/api/setup/checkExistingInvites/route.tsx
@@ -19,12 +19,16 @@ export async function GET(_req: NextRequest) {
 
     if (!invites.length) {
       return NextResponse.json(
-        { message: "No invitations found." },
-        { status: 404 }
+      { message: "No invitations found." },
+      { status: 400 }
       );
     }
-
-    return NextResponse.json({ invites }, { status: 200 });
+    const spaceName = await prisma.space.findUnique({
+      where: { id: invites[0].spaceId },
+      select: { name: true },
+    })
+    const name = spaceName?.name
+    return NextResponse.json({ invites,name }, { status: 200 });
   } catch (error) {
     console.error("Error fetching invitations:", error);
     return NextResponse.json(
@@ -35,12 +39,13 @@ export async function GET(_req: NextRequest) {
 }
 
 
-export async function POST(req: NextRequest) {
+export async function POST(req: NextRequest, res: NextResponse) {
   try {
     const { success, user, error, status } = await getUserFromSession();
     if (!success || !user) {
       return NextResponse.json({ error }, { status });
     }
+
     const body = await req.json();
     const { requestId } = body;
 
@@ -51,19 +56,69 @@ export async function POST(req: NextRequest) {
       );
     }
 
+    const joinRequest = await prisma.invitation.findUnique({
+      where: { id: requestId },
+      select: {
+        id: true,
+        spaceId: true,
+        email: true,
+        role: true,
+      },
+    });
+
+    if (!joinRequest) {
+      return NextResponse.json(
+        { message: "Invalid Join Request." },
+        { status: 400 }
+      );
+    }
+
+    const newMember = await prisma.spaceMember.create({
+      data: {
+        joinedAt: new Date(),
+        role: joinRequest.role,
+        email: joinRequest.email,
+        spaceId: joinRequest.spaceId,
+        userId: user.id,
+      },
+    });
+
+    if (!newMember) {
+      return NextResponse.json(
+        { message: "Failed to accept invitation." },
+        { status: 500 }
+      );
+    }
+
     const deletedRequest = await prisma.invitation.delete({
       where: { id: requestId },
     });
 
-    return NextResponse.json(deletedRequest, { status: 200 });
-  } catch (error) {
-    console.error("Error deleting invitation:", error);
+    if (!deletedRequest) {
+      return NextResponse.json(
+        { message: "Failed to delete the join request." },
+        { status: 500 }
+      );
+    }
+
     return NextResponse.json(
-      { message: "An internal server error occurred." },
+      { message: "Invitation accepted." },
+      { status: 200 }
+    );
+  } catch (error: any) {
+    console.error("Error processing join request:", error);
+
+    return NextResponse.json(
+      {
+        message:
+          "An unexpected error occurred while processing the accept request.",
+        error: error.message || "Internal Server Error",
+      },
       { status: 500 }
     );
   }
 }
+
 export async function PUT(req: NextRequest, _res: NextResponse) {
   try {
     const { success, user, error, status } = await getUserFromSession();
